+++ examples/contrastive-image-text/clip_media_pipe.py
@@ -1,15 +1,33 @@
-# Copyright (C) 2023 Habana Labs, Ltd. an Intel Company
+# coding=utf-8
+# Copyright 2023 The HuggingFace Team All rights reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 
 import numpy as np
-from habana_frameworks.mediapipe import fn
-from habana_frameworks.mediapipe.backend.nodes import opnode_tensor_info
-from habana_frameworks.mediapipe.backend.operator_specs import schema
-from habana_frameworks.mediapipe.media_types import dtype, ftype, imgtype, randomCropType, readerOutType
-from habana_frameworks.mediapipe.mediapipe import MediaPipe
-from habana_frameworks.mediapipe.operators.media_nodes import MediaReaderNode
-from habana_frameworks.mediapipe.operators.reader_nodes.read_image_from_dir import get_max_file
-from habana_frameworks.torch.hpu import get_device_name
 from torch.utils.data.sampler import BatchSampler
+
+
+try:
+    from habana_frameworks.mediapipe import fn
+    from habana_frameworks.mediapipe.backend.nodes import opnode_tensor_info
+    from habana_frameworks.mediapipe.backend.operator_specs import schema
+    from habana_frameworks.mediapipe.media_types import dtype, ftype, imgtype, randomCropType, readerOutType
+    from habana_frameworks.mediapipe.mediapipe import MediaPipe
+    from habana_frameworks.mediapipe.operators.media_nodes import MediaReaderNode
+    from habana_frameworks.mediapipe.operators.reader_nodes.read_image_from_dir import get_max_file
+    from habana_frameworks.torch.hpu import get_device_name
+except ImportError:
+    pass
 
 
 class read_image_text_from_dataset(MediaReaderNode):
@@ -105,16 +123,19 @@ setattr(fn, "ClipDataReader", op_class)
 
 class ClipMediaPipe(MediaPipe):
     """
-    Class defining clip media pipe.
+    Class defining clip media pipe:
+        read data --> image decoding (include crop and resize) --> crop mirror normalize
+
+    The image transformation in origin pytorch transformation:
+        aspect ratio preserving resize -> center crop -> normalize
 
     """
 
     batch_sampler = None
     instance_count = 0
 
-    def __init__(self, is_training=True, dataset=None, sampler=None, batch_size=512, drop_last=False, queue_depth=1):
+    def __init__(self, dataset=None, sampler=None, batch_size=512, drop_last=False, queue_depth=1):
         self.device = get_device_name()
-        self.is_training = is_training
         self.dataset = dataset
         self.drop_last = drop_last
         self.sampler = sampler
@@ -128,13 +149,9 @@ class ClipMediaPipe(MediaPipe):
             device=self.device, batch_size=batch_size, prefetch_depth=queue_depth, pipe_name=pipe_name
         )
 
-        ClipMediaPipe.instance_count += 1
-
-    def definegraph(self):
-        res_pp_filter = ftype.BICUBIC
         self.input = fn.ClipDataReader(label_dtype=dtype.UINT32, dataset=self.dataset)
-        jpegs, input_ids, attention_masks = self.input()
         def_output_image_size = [self.image_size, self.image_size]
+        res_pp_filter = ftype.BICUBIC
         self.decode = fn.ImageDecoder(
             device=self.device,
             output_format=imgtype.RGB_P,
@@ -142,13 +159,13 @@ class ClipMediaPipe(MediaPipe):
             resize=def_output_image_size,
             resampling_mode=res_pp_filter,
         )
-        images = self.decode(jpegs)
+
         cmn_pos_offset = 0.5
         normalize_mean = np.array(self.dataset.image_mean).astype(np.float32) * 255
         normalize_std = 1 / (np.array(self.dataset.image_std).astype(np.float32) * 255)
         norm_mean = fn.MediaConst(data=normalize_mean, shape=[1, 1, 3], dtype=dtype.FLOAT32)
         norm_std = fn.MediaConst(data=normalize_std, shape=[1, 1, 3], dtype=dtype.FLOAT32)
-        cmn = fn.CropMirrorNorm(
+        self.cmn = fn.CropMirrorNorm(
             crop_w=self.image_size,
             crop_h=self.image_size,
             crop_pos_x=cmn_pos_offset,
@@ -156,7 +173,13 @@ class ClipMediaPipe(MediaPipe):
             crop_d=0,
             dtype=dtype.FLOAT32,
         )
-        mean = norm_mean()
-        std = norm_std()
-        images = cmn(images, mean, std)
+        self.mean = norm_mean()
+        self.std = norm_std()
+
+        ClipMediaPipe.instance_count += 1
+
+    def definegraph(self):
+        jpegs, input_ids, attention_masks = self.input()
+        images = self.decode(jpegs)
+        images = self.cmn(images, self.mean, self.std)
         return images, input_ids, attention_masks
