+++ examples/contrastive-image-text/run_clip.py
@@ -31,6 +31,7 @@ from typing import Optional
 import torch
 import transformers
 from datasets import load_dataset
+from habana_dataloader_trainer import HabanaDataloaderTrainer
 from PIL import Image
 from torchvision.io import ImageReadMode, read_image
 from torchvision.transforms import CenterCrop, ConvertImageDtype, Normalize, Resize
@@ -41,8 +42,7 @@ from transformers import (
     AutoTokenizer,
     HfArgumentParser,
 )
-from transformers.trainer_pt_utils import IterableDatasetShard
-from transformers.trainer_utils import get_last_checkpoint, seed_worker
+from transformers.trainer_utils import get_last_checkpoint
 from transformers.utils import check_min_version, send_example_telemetry
 from transformers.utils.versions import require_version
 
@@ -56,132 +56,6 @@ logger = logging.getLogger(__name__)
 check_min_version("4.28.0")
 
 require_version("datasets>=1.8.0", "To fix: pip install -r examples/pytorch/contrastive-image-text/requirements.txt")
-
-
-class MediaApiDataLoader(torch.utils.data.DataLoader):
-    def __init__(
-        self,
-        dataset,
-        batch_size=1,
-        sampler=None,
-        collate_fn=None,
-        drop_last=False,
-        num_workers=0,
-        pin_memory=False,
-        worker_init_fn=None,
-    ):
-        self.dataset = dataset
-        self.sampler = sampler
-        self.fallback_activated = False
-
-        try:
-            from clip_media_pipe import ClipMediaPipe
-            from habana_frameworks.mediapipe.plugins.iterator_pytorch import HPUGenericPytorchIterator
-
-            pipeline = ClipMediaPipe(
-                is_training=True,
-                dataset=dataset,
-                sampler=sampler,
-                batch_size=batch_size,
-                drop_last=drop_last,
-                queue_depth=3,
-            )
-            self.iterator = HPUGenericPytorchIterator(mediapipe=pipeline)
-        except Exception as e:
-            print("Warning: using Pytorch native dataloader. ", e)
-            self.fallback_activated = True
-            dataset.set_transform(dataset.transform_func)
-            super(MediaApiDataLoader, self).__init__(
-                dataset,
-                batch_size=batch_size,
-                sampler=sampler,
-                collate_fn=collate_fn,
-                drop_last=drop_last,
-                num_workers=num_workers,
-                pin_memory=pin_memory,
-                worker_init_fn=worker_init_fn,
-            )
-
-    def __len__(self):
-        if self.fallback_activated:
-            return super().__len__()
-        else:
-            return len(self.iterator)
-
-    def __iter__(self):
-        if self.fallback_activated:
-            return super().__iter__()
-        else:
-            self.iterator.__iter__()
-        return self
-
-    def __next__(self):
-        if self.fallback_activated:
-            return super().__next__()
-
-        data = next(self.iterator)
-        return {
-            "pixel_values": data[0],
-            "input_ids": data[1],
-            "attention_mask": data[2],
-            "return_loss": True,
-        }
-
-
-class HabanaDataloaderTrainer(GaudiTrainer):
-    def get_train_dataloader(self):
-        """
-        Returns the training [`~torch.utils.data.DataLoader`].
-
-        Will use no sampler if `train_dataset` does not implement `__len__`, a random sampler (adapted to distributed
-        training if necessary) otherwise.
-
-        Subclass and override this method if you want to inject some custom behavior.
-        """
-        if self.train_dataset is None:
-            raise ValueError("Trainer: training requires a train_dataset.")
-
-        train_dataset = self.train_dataset
-        data_collator = self.data_collator
-        from transformers.utils import is_datasets_available
-
-        if is_datasets_available():
-            import datasets
-        if is_datasets_available() and isinstance(train_dataset, datasets.Dataset):
-            train_dataset = self._remove_unused_columns(train_dataset, description="training")
-        else:
-            data_collator = self._get_collator_with_removed_columns(data_collator, description="training")
-
-        if isinstance(train_dataset, torch.utils.data.IterableDataset):
-            if self.args.world_size > 1:
-                train_dataset = IterableDatasetShard(
-                    train_dataset,
-                    batch_size=self._train_batch_size,
-                    drop_last=self.args.dataloader_drop_last,
-                    num_processes=self.args.world_size,
-                    process_index=self.args.process_index,
-                )
-
-            return MediaApiDataLoader(
-                train_dataset,
-                batch_size=self._train_batch_size,
-                collate_fn=data_collator,
-                num_workers=self.args.dataloader_num_workers,
-                pin_memory=self.args.dataloader_pin_memory,
-            )
-
-        train_sampler = self._get_train_sampler()
-
-        return MediaApiDataLoader(
-            train_dataset,
-            batch_size=self._train_batch_size,
-            sampler=train_sampler,
-            collate_fn=data_collator,
-            drop_last=self.args.dataloader_drop_last,
-            num_workers=self.args.dataloader_num_workers,
-            pin_memory=self.args.dataloader_pin_memory,
-            worker_init_fn=seed_worker,
-        )
 
 
 @dataclass
